UK_ACCIDENTDATA_CLI/accident_cli/requirements.txt
mysql-connector-python
sqlalchemy
prettytable
folium